moved files around, added game mode enum
