+++ src/enums/GameMode.ts
@@ -1,0 +1,5 @@
+export enum GameMode {
+    SinglePlayer = 1,
+    MultiPlayer = 2,
+    
+}
