+++ src/interfaces/ScoringStrategy.ts
